requirements.txt
amqp==5.1.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
biopython==1.79
Brotli==1.0.9
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Compress==1.13
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
plotly==5.10.0
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.1.0
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1